refactor(modules/fn;module/object;test): fix module conflict
BREAKING CHANGE: 模块下文件名即是导出的方法名,下划线开始的方法名无法以 api 的形式访问
