+++ src/modules/fn/_isArray.js
@@ -1,0 +1,4 @@
+const _typeof = require('./_typeof')
+const _isArray = (target) => _typeof(target) === 'Array'
+
+module.exports = _isArray
