+++ src/modules/fn/_isObject.js
@@ -1,0 +1,4 @@
+const _typeof = require('./_typeof')
+const _isObject = (target) => _typeof(target) === 'Object'
+
+module.exports = _isObject
