+++ src/modules/fn/_typeof.js
@@ -1,0 +1,3 @@
+const _typeof = (target) => Object.prototype.toString.call(target).match(/\[object\s(\w+)\]/)[1]
+
+module.exports = _typeof
