+++ src/modules/fn/debounce.js
@@ -1,0 +1,13 @@
+function debounce(fn, delay) {
+  let timer = null
+  return (...args) => {
+    if (timer) {
+      clearTimeout(timer)
+    }
+    timer = setTimeout(() => {
+      fn.apply(this, args)
+    }, delay)
+  }
+}
+
+module.exports = debounce
